fwk: implement the battery extender feature
cherry pick Google CLs for battery exceeded
list:
[Update EC_CMD_CHARGE_CONTROL to version 2](https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2929340)
[chgstv2: Add unit test for battery sustainer](https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2987734/4)
[chgstv2: Add battery sustainer](https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2929347)
[chgstv2: Refactor charger_discharge_on_ac](https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2986848)
manual add CLs
[charge_state: Add EC_CMD_CHARGE_CONTROL V3](https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4824703)

fwk CLs:
[fwk: override the display charge value for Windows system](https://github.com/FrameworkComputer/ec/pull/1042)

The purpose of this function is to automatically drop the battery max charge voltage if the system is left plugged into AC for a long time to preserve battery life.
When the timer is exceeded the battery charge voltage should be reduced to 4.35V/cell. Two additional days after the timer has exceeded, the battery charge voltage shall be reduced to 4.3V/cell.
The timer should reset to 0 days any time the system is in S0/S0ix and not attached to AC for 30 minutes or longer.

BRANCH=fwk-hx20-hx30-4410
BUG=https://app.clickup.com/t/86eq06zen
TEST=zmake build hx30 pass
TEST=trigger extender timer check chg ctl mode is discharge(2) battery also has discharge
after battery percentage trace to upper value  chg ctl mode change to IDLE

Signed-off-by: LeoCX_Tsai <LeoCX_Tsai@compal.com>